Bugfix completed

+++ tools/store/store.h
@@ -16,7 +16,7 @@ uint16_t Return_16bit(uint8_t addr);
 void Store_ID(uint16_t id);
 uint16_t Return_ID();
 void Store_Address(unit16_t address, uint8_t media);
-uint8_t Return_Addresses(uint8_t media);
+uint16_t Return_Addresses(uint8_t media);
 void Store_PeerAddresses(uint16_t *addresses, uint8_t n_addresses);
 void Return_PeerAddresses(uint16_t *addresses, uint8_t n_addresses);
 uint16_t Return_SinglePeerAddresses(uint8_t n_addr);
